websockets: move source from experimental to k6/websockets
Move the websockets module source files from
internal/js/modules/k6/experimental/websockets/ to
internal/js/modules/k6/websockets/.

This simplifies future maintenance by keeping the canonical
source in the stable module location.
