+++ internal/js/jsmodules.go
@@ -15,11 +15,11 @@ import (
 	"go.k6.io/k6/internal/js/modules/k6/experimental/csv"
 	"go.k6.io/k6/internal/js/modules/k6/experimental/fs"
 	"go.k6.io/k6/internal/js/modules/k6/experimental/streams"
-	expws "go.k6.io/k6/internal/js/modules/k6/experimental/websockets"
 	"go.k6.io/k6/internal/js/modules/k6/grpc"
 	"go.k6.io/k6/internal/js/modules/k6/metrics"
 	"go.k6.io/k6/internal/js/modules/k6/secrets"
 	"go.k6.io/k6/internal/js/modules/k6/timers"
+	"go.k6.io/k6/internal/js/modules/k6/websockets"
 	"go.k6.io/k6/internal/js/modules/k6/ws"
 	"go.k6.io/k6/js/common"
 	"go.k6.io/k6/js/modules"
@@ -45,7 +45,7 @@ func getInternalJSModules() map[string]interface{} {
 		"k6/metrics":     metrics.New(),
 		"k6/secrets":     secrets.New(),
 		"k6/timers":      timers.New(),
-		"k6/websockets":  expws.New(),
+		"k6/websockets":  websockets.New(),
 		"k6/ws":          ws.New(),
 
 		// Experimental modules
@@ -54,7 +54,7 @@ func getInternalJSModules() map[string]interface{} {
 		"k6/experimental/streams": streams.New(),
 
 		// Deprecated modules
-		"k6/experimental/websockets": newWarnExperimentalModule(expws.New(),
+		"k6/experimental/websockets": newWarnExperimentalModule(websockets.New(),
 			"k6/experimental/websockets is deprecated and will be removed in a future release."+
 				" Please use k6/websockets instead."),
 		"k6/experimental/redis": newWarnExperimentalModule(redis.New(),
