+++ internal/js/modules/k6/websockets/autobahn_tests/script.js
