+++ internal/js/modules/k6/websockets/blob.go
