+++ internal/js/modules/k6/websockets/events/events.go
