+++ internal/js/modules/k6/websockets/helpers.go
