+++ internal/js/modules/k6/websockets/main_test.go
