+++ internal/js/modules/k6/websockets/params.go
